Task/implementing add node

+++ app/org/org-node-detail/org-node-detail.component.ts
@@ -13,8 +13,10 @@ import { OrgNodeModel, OrgService } from "../shared/index";
 
 export class OrgNodeDetailComponent {
     @Input() selectedOrgNode: OrgNodeModel;
+    @Input() isAddMode: boolean;
     @Output() deleteNode = new EventEmitter<OrgNodeModel>();
     @Output() updateNode = new EventEmitter<OrgNodeModel>();
+    @Output() addNode = new EventEmitter<OrgNodeModel>();
 
     private isEditMode: boolean;
     private editNodeDetails: OrgNodeModel;
@@ -30,17 +32,58 @@ export class OrgNodeDetailComponent {
         return (node.children != null);
     }
 
+    private isNullOrEmpty(value: string) {
+        if (value && value.trim().length > 0) {
+            return false;
+        }
+        return true;
+    }
+
     private onSubmit(form: NgForm) {
-        let data = JSON.stringify(form.value, null, 2);
-        this.editNodeDetails = new OrgNodeModel();
-        this.editNodeDetails.NodeFirstName = form.value.firstName;
-        this.editNodeDetails.NodeLastName = form.value.lastName;
-        this.editNodeDetails.Description = form.value.description;
-        this.editNodeDetails.children = this.selectedOrgNode.children;
-        this.editNodeDetails.NodeID = this.selectedOrgNode.NodeID;
-        this.editNodeDetails.OrgID = this.selectedOrgNode.OrgID;
-        this.editNodeDetails.ParentNodeID = this.selectedOrgNode.ParentNodeID;
-        this.editNode(this.editNodeDetails);
+        if (!this.isNullOrEmpty(form.value.firstName)) {
+            this.editNodeDetails = new OrgNodeModel();
+            this.editNodeDetails.NodeFirstName = form.value.firstName;
+            this.editNodeDetails.NodeLastName = form.value.lastName;
+            this.editNodeDetails.Description = form.value.description;
+            this.editNodeDetails.children = this.selectedOrgNode.children;
+            this.editNodeDetails.NodeID = this.selectedOrgNode.NodeID;
+            this.editNodeDetails.OrgID = this.selectedOrgNode.OrgID;
+            this.editNodeDetails.ParentNodeID = this.selectedOrgNode.ParentNodeID;
+
+            this.editNode(this.editNodeDetails);
+        }
+    }
+
+    private onFirstNameBlurred(fname: string) {
+        if (this.isAddMode) {
+            if (!this.isNullOrEmpty(fname)) {
+                let node = new OrgNodeModel();
+                node.NodeFirstName = fname;
+                node.children = this.selectedOrgNode.children;
+                node.OrgID = this.selectedOrgNode.OrgID;
+                node.ParentNodeID = this.selectedOrgNode.ParentNodeID;
+
+                this.addNewNode(node);
+            }
+        }
+    }
+
+    private emitaddNodeNotification(data: OrgNodeModel) {
+        if (data) {
+            this.addNode.emit(data);
+            this.isEditMode = true;
+            this.editNodeDetails = null;
+        }
+    }
+
+    private addNewNode(node: OrgNodeModel) {
+        if (!node) { return; }
+        // we don"t really need to send any child info to the server at this point
+        node.children = null;
+        this.orgService.addNode(node)
+            .subscribe(data => this.emitaddNodeNotification(data),
+            error => this.handleError(error),
+            () => console.log("Node Added Complete"));
     }
 
     private editNode(node: OrgNodeModel) {
@@ -62,8 +105,11 @@ export class OrgNodeDetailComponent {
     }
 
     private onDeleteNodeClicked() {
-        if (this.selectedOrgNode) {
-            if (this.selectedOrgNode.children == null) {
+        if (this.isAddMode) {
+            this.deleteNode.emit(this.selectedOrgNode);
+        }
+        else {
+            if (this.selectedOrgNode.children == null && !this.isAddMode) {
                 this.orgService.deleteNode(this.selectedOrgNode.NodeID)
                     .subscribe(data => this.emitDeleteNodeNotification(data),
                     error => this.handleError(error),
@@ -90,7 +136,12 @@ export class OrgNodeDetailComponent {
     }
 
     private handleError(err) {
-        alert("OOPs!! Something went wrong!! ");
+        try {
+            let errorMessage = JSON.parse(err._body);
+            alert(errorMessage.Message);
+        } catch (ex) {
+            alert("OOPs!! Something went wrong!! ");
+        }
         console.log(err);
         this.isEditMode = false;
         this.editNodeDetails = null;
